# libs/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@daytona.com
    name: Daytona Platforms Inc.
    url: https://www.daytona.io
  description: Daytona AI platform API Docs
  title: Daytona
  version: '1.0'
servers:
  - url: http://localhost:3000
paths:
...
  /workspace/{workspaceId}/build-logs:
    get:
      operationId: getBuildLogs
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: ID of the workspace
          explode: false
          in: path
          name: workspaceId
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          description: Build logs stream
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getBuildLogs_400_response'
          description: Bad request
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getBuildLogs_400_response'
          description: Build logs not found
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Get build logs
      tags:
        - workspace
...
  /images/{id}:
    delete:
      operationId: removeImage
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Image ID
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          description: Image has been deleted
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Delete image
      tags:
        - images
    get:
      operationId: getImage
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
        - description: Image ID
          explode: false
          in: path
          name: id
          required: true
          schema:
            type: string
          style: simple
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageDto'
          description: The image
        '404':
          description: Image not found
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Get image by ID
      tags:
        - images
  /images/build:
    post:
      operationId: buildImage
      parameters:
        - description: Use with JWT to specify the organization ID
          explode: false
          in: header
          name: X-Daytona-Organization-ID
          required: false
          schema:
            type: string
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BuildImage'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageDto'
          description: The image has been successfully built.
        '400':
          description: Bad request - Invalid build parameters
      security:
        - bearer: []
        - oauth2:
            - openid
            - profile
            - email
      summary: Build a Docker image
      tags:
        - images
...
components:
  schemas:
...
    CreateBuildInfo:
      properties:
        dockerfileContent:
          description: The Dockerfile content used for the build
          example: |-
            FROM node:14
            WORKDIR /app
            COPY . .
            RUN npm install
            CMD ["npm", "start"]
          type: string
        contextHashes:
          description: The context hashes used for the build
          example:
            - hash1
            - hash2
          items:
            type: string
          type: array
      required:
        - dockerfileContent
      type: object
...
    ImageState:
      enum:
        - build_pending
        - building
        - pending
        - pulling_image
        - pending_validation
        - validating
        - active
        - error
        - removing
      type: string
    ImageDto:
      example:
        organizationId: organizationId
        general: true
        createdAt: 2000-01-23T04:56:07.000+00:00
        size: 0.8008281904610115
        lastUsedAt: 2000-01-23T04:56:07.000+00:00
        entrypoint:
          - entrypoint
          - entrypoint
        errorReason: errorReason
        name: name
        id: id
        state: ''
        enabled: true
        updatedAt: 2000-01-23T04:56:07.000+00:00
      properties:
        id:
          type: string
        organizationId:
          type: string
        general:
          type: boolean
        name:
          type: string
        enabled:
          type: boolean
        state:
          allOf:
            - $ref: '#/components/schemas/ImageState'
        size:
          nullable: true
          type: number
        entrypoint:
          items:
            type: string
          nullable: true
          type: array
        errorReason:
          nullable: true
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        lastUsedAt:
          format: date-time
          nullable: true
          type: string
      required:
        - createdAt
        - enabled
        - entrypoint
        - errorReason
        - general
        - id
        - lastUsedAt
        - name
        - size
        - state
        - updatedAt
      type: object
    BuildImage:
      example:
        buildInfo: ''
        name: my-custom-image:1.0.0
      properties:
        name:
          description: The name of the image to build
          example: my-custom-image:1.0.0
          type: string
        buildInfo:
          allOf:
            - $ref: '#/components/schemas/CreateBuildInfo'
          description: Build information for the image
      required:
        - buildInfo
        - name
      type: object
...
    getBuildLogs_400_response:
      example:
        message: message
        statusCode: 0.8008281904610115
      properties:
        message:
          type: string
        statusCode:
          type: number
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http
    oauth2:
      openIdConnectUrl: http://localhost:3000/.well-known/openid-configuration
      type: openIdConnect